pyproject.toml
[project]
name = "accesscheck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "easyocr>=1.7.2",
    "openai==0.28",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.6",
    "pillow>=11.1.0",
    "pre-commit>=4.2.0",
    "requests>=2.32.3",
    "ruff>=0.11.5",
    "selenium>=4.31.0",
    "tkcalendar>=1.6.1",
]
